JAMES-2521 Copier shouldhave its own tool module
