+++ mailbox/spring/pom.xml
@@ -73,7 +73,7 @@
         </dependency>
         <dependency>
             <groupId>${james.groupId}</groupId>
-            <artifactId>apache-james-mailbox-tool</artifactId>
+            <artifactId>apache-james-mailbox-tools-copier</artifactId>
         </dependency>
         <dependency>
             <groupId>${james.groupId}</groupId>
