+++ mailbox/tools/copier/pom.xml
@@ -21,15 +21,13 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
+        <artifactId>apache-james-mailbox-tools</artifactId>
         <groupId>org.apache.james</groupId>
-        <artifactId>apache-james-mailbox</artifactId>
         <version>3.2.0-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>apache-james-mailbox-tool</artifactId>
-    <packaging>bundle</packaging>
-    <name>Apache James :: Mailbox :: Tools</name>
+    <artifactId>apache-james-mailbox-tools-copier</artifactId>
+    <name>Apache James :: Mailbox :: Tools :: Copier</name>
 
     <dependencies>
         <dependency>
@@ -75,6 +73,7 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
+            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.assertj</groupId>
@@ -91,4 +90,5 @@
             <artifactId>slf4j-api</artifactId>
         </dependency>
     </dependencies>
+
 </project>
