+++ server/container/guice/jmx/src/main/java/org/apache/james/modules/server/JMXServerModule.java
@@ -36,8 +36,8 @@ import org.apache.james.domainlist.api.DomainListManagementMBean;
 import org.apache.james.domainlist.lib.DomainListManagement;
 import org.apache.james.lifecycle.api.Configurable;
 import org.apache.james.mailbox.copier.MailboxCopier;
-import org.apache.james.mailbox.copier.MailboxCopierImpl;
 import org.apache.james.mailbox.indexer.ReIndexer;
+import org.apache.james.mailbox.tools.copier.MailboxCopierImpl;
 import org.apache.james.mailetcontainer.api.jmx.MailSpoolerMBean;
 import org.apache.james.mailetcontainer.impl.JamesMailSpooler;
 import org.apache.james.rrt.api.RecipientRewriteTableManagementMBean;
